Refactor evaluate view components into a separate folder
- Add changelog entry
